[BugFix] 근접 몬스터 공격 버그 픽스, 포탑 Hurt 처리 완료, Core HP UI 버그 픽스, Player BuildSystem 천장 설치 버그 수정

+++ Assets/Turret/Scripts/Core.cs
@@ -162,7 +162,7 @@ public class Core : MonoBehaviour
     public void Hurt(int damge)
     {
         nowHp -= damge;
-        CoreHPBar.fillAmount = nowHp / maxHp;
+        CoreHPBar.fillAmount = (float)((float)nowHp / (float)maxHp);
         CoreHPText.text = $"{nowHp} / {maxHp}";
     }
